requirements.txt
mega.py
aiogram
pydantic-settings
aiosqlite
sqlalchemy